fix: scrolling, header flicker, and window drag flicker bugs
1. Scrolling fix (mouse_scroll_delta type mismatch):
   - mouse.c: Changed mouse_scroll_delta from int8_t to int to match
     the extern declaration in mouse.h and all consumers. The type
     mismatch caused memory corruption: reading 4 bytes from a 1-byte
     variable returned garbage, and writing 0 via 'extern int' cleared
     4 bytes, zeroing adjacent mouse_has_wheel and disabling scroll.
   - mouse.c: Changed mouse_scroll_delta assignment from overwrite (=)
     to accumulate (+=) so rapid scroll events between frames are not
     lost. The value is consumed and cleared each frame by
     sys_mouse_scroll().
   - mouse.h: Removed incorrect 'extern int mouse_has_wheel' (it's
     static in mouse.c, not exported).
   - api.c: Updated extern to 'int mouse_scroll_delta' to match.

2. Header flicker fix (context menu bounds + dropdown overlap):
   - bubbleview.c ctx_menu_show(): Added x<0 and y<0 bounds clamping.
     Previously, right-clicking near the top of the screen could set
     the context menu's y-coordinate negative, causing visual
     corruption and flicker as the menu drew outside screen bounds.
   - bubbleview.c process_global_bar(): Added suppress_dropdowns flag
     when context menu is active. Prevents the menu bar's dropdown
     menus from drawing simultaneously with the context menu, which
     caused overlapping pop-ups fighting for the same screen region.

3. Window drag flicker fixes:
   - bubbleview.c: Added drag_just_released flag for one final
     dirty-region frame on mouse release, instead of falling through
     to a slower full-screen redraw.
   - desktop.c desktop_draw_icons(): Each icon now repaints its
     background from the wallpaper cache before drawing, preventing
     stale highlights or ghost pixels from the previous frame when
     using the dirty-region optimisation during window dragging.

+++ sys/api.c
@@ -138,11 +138,11 @@ int sys_mouse_read(int* x, int* y, int* left_click) {
     return (mouse_btn_left | (mouse_btn_right << 1));
 }
 
-extern int8_t mouse_scroll_delta;
+extern int mouse_scroll_delta;
 int sys_mouse_scroll() {
-    int8_t delta = mouse_scroll_delta;
+    int delta = mouse_scroll_delta;
     mouse_scroll_delta = 0;  // Consume the scroll event
-    return (int)delta;
+    return delta;
 }
 
 void sys_print(const char* str) { vga_print(str); }
